% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_url_elements}
\alias{get_url_elements}
\alias{get_scheme}
\alias{get_host}
\alias{get_port}
\alias{get_query}
\alias{get_path}
\alias{get_fragment}
\alias{get_username}
\alias{get_password}
\title{Get URL elements}
\usage{
get_scheme(url)

get_host(url)

get_port(url)

get_query(url)

get_path(url)

get_fragment(url)

get_username(url)

get_password(url)
}
\arguments{
\item{url}{character vector of urls}
}
\value{
a vector of characters
}
\description{
Get URL elements
}
\examples{
library(weburl)

url <- "https://user:pass@example.com:1234/foo/bar?baz#quux"

get_scheme(url)
get_host(url)
get_port(url)
get_path(url)
get_query(url)
get_fragment(url)
get_username(url)
get_password(url)
}
