% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_scheme}
\alias{set_scheme}
\alias{set_host}
\alias{set_port}
\alias{set_path}
\alias{set_query}
\alias{set_fragment}
\alias{set_username}
\alias{set_password}
\title{Set URL elements}
\usage{
set_scheme(url, scheme)

set_host(url, host)

set_port(url, port)

set_path(url, path)

set_query(url, query)

set_fragment(url, fragment)

set_username(url, username)

set_password(url, password)
}
\arguments{
\item{url}{character vector of urls}
}
\value{
a vector of characters
}
\description{
Set URL elements
}
