% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{url_encode}
\alias{url_encode}
\title{Encode Url}
\usage{
url_encode(url)
}
\arguments{
\item{url}{character vector of urls}
}
\value{
a vector of characters
}
\description{
Encode Url
}
\examples{
library(weburl)

url <- "https:www.google.com/dummy=path yo"

url_encode(url)
}
