% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{url_parse}
\alias{url_parse}
\title{Parse url}
\usage{
url_parse(url)
}
\arguments{
\item{url}{character vector of urls}
}
\value{
a data.frame consisting of the columns scheme, host, port, path, query and fragment.
}
\description{
Parse url
}
