% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{weburl-package}
\alias{weburl}
\alias{weburl-package}
\title{weburl: Parse Urls}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
